+++ Sources/Primitive.swift
@@ -2,7 +2,7 @@ import Foundation
 
 public enum Primitive {
     case date(Date)
-    case float(Float)
+    case double(Double)
     case int(Int)
     case null
     case string(String)
